# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "IMBiz Swagger API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:18181",
    "basePath": "/",
    "paths": {
...
        "/mysql/create": {
            "post": {
                "description": "POST Mysql Create",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "POST Mysql Create",
                "parameters": [
                    {
                        "description": "text",
                        "name": "TEXT",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mysql/delete": {
            "delete": {
                "description": "DELETE Mysql Delete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "DELETE Mysql Delete",
                "parameters": [
                    {
                        "description": "index",
                        "name": "INDEX",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/mysql/update": {
            "put": {
                "description": "PUT Mysql Update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "PUT Mysql Update",
                "parameters": [
                    {
                        "description": "text",
                        "name": "TEXT",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "index",
                        "name": "INDEX",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}